refactor: revert previous changes
