--- packages/trader/src/AppV2/Components/FocusedInput/focused-input.scss
@@ -1,5 +1,0 @@
-.input {
-    &--focused {
-        position: absolute;
-    }
-}
