--- packages/trader/src/AppV2/Components/FocusedInput/index.ts
@@ -1,4 +1,0 @@
-import FocusedInput from './focused-input';
-import './focused-input.scss';
-
-export default FocusedInput;
